+++ ddataflow/setup_project.py
@@ -2,16 +2,16 @@ def setup_project(config_file):
     content = """
 from ddataflow import DDataflow
 
-config = {
-    # (optional) add here tables with ddataflow predefined sampling default sampling (a good way to start)
-    # files with default_sampling are not supported at the moment
-    "sources_with_default_sampling": [],
+_config = {
     # add here your tables or paths with customized sampling logic
-    "data_sources": {},
+    "data_sources": {
+        'demo_tours': {
+            'source': lambda spark: spark.table('demo_tours'),
+            'default_sampling': True,
+        }
+    },
     # add here your writing logic
     "data_writers": {},
-    # this is the name of the project to identify this project in the filesystem
-    # do not use - minus - signs in the name
     "project_folder_name": "my_project",
     # to customize the location of your datasets 
     # "snapshot_path": "dbfs:/another_databricks_path",